+++ consensus/reactor.go
@@ -1712,6 +1712,9 @@ func (m *NewValidBlockMessage) ValidateBasic() error {
 	if err := m.BlockPartSetHeader.ValidateBasic(); err != nil {
 		return cmterrors.ErrWrongField{Field: "BlockPartSetHeader", Err: err}
 	}
+	if err := m.BlockParts.ValidateBasic(); err != nil {
+		return cmterrors.ErrWrongField{Field: "BlockParts", Err: err}
+	}
 	if m.BlockParts.Size() == 0 {
 		return cmterrors.ErrRequiredField{Field: "blockParts"}
 	}
@@ -1771,6 +1774,9 @@ func (m *ProposalPOLMessage) ValidateBasic() error {
 	}
 	if m.ProposalPOLRound < 0 {
 		return cmterrors.ErrNegativeField{Field: "ProposalPOLRound"}
+	}
+	if err := m.ProposalPOL.ValidateBasic(); err != nil {
+		return cmterrors.ErrWrongField{Field: "ProposalPOL", Err: err}
 	}
 	if m.ProposalPOL.Size() == 0 {
 		return cmterrors.ErrRequiredField{Field: "ProposalPOL"}
@@ -1916,6 +1922,9 @@ func (m *VoteSetBitsMessage) ValidateBasic() error {
 	}
 	if err := m.BlockID.ValidateBasic(); err != nil {
 		return cmterrors.ErrWrongField{Field: "BlockID", Err: err}
+	}
+	if err := m.Votes.ValidateBasic(); err != nil {
+		return cmterrors.ErrWrongField{Field: "Votes", Err: err}
 	}
 	// NOTE: Votes.Size() can be zero if the node does not have any
 	if m.Votes.Size() > types.MaxVotesCount {