+++ consensus/wal_test.go
@@ -91,7 +91,6 @@ func TestWALEncoderDecoder(t *testing.T) {
 	b := new(bytes.Buffer)
 
 	for _, msg := range msgs {
-		msg := msg
 
 		b.Reset()